blogging-platform/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.0
django-filter>=23.0
python-decouple>=3.8
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.3.0
